Move the functions in classify_intersections.rs to road.rs and intersection.rs
