+++ osm2streets/src/transform/collapse_short_road.rs
@@ -140,7 +140,7 @@ impl StreetNetwork {
 
         // We just connected a bunch of things to keep_i. Fix ordering and movements.
         self.sort_roads(keep_i);
-        self.recalculate_movements(keep_i);
+        self.update_movements(keep_i);
 
         // If we're deleting the target of a simple restriction somewhere, update it.
         for road in self.roads.values_mut() {
